Make index.js files just export files
