+++ packages/shared-src/src/models/fhir/ServiceRequest/index.js
@@ -1,107 +1,1 @@
-import { DataTypes } from 'sequelize';
-
-import { FhirResource } from '../Resource';
-
-import {
-  FHIR_INTERACTIONS,
-  FHIR_SEARCH_PARAMETERS,
-  FHIR_SEARCH_TOKEN_TYPES,
-} from '../../../constants';
-import { getQueryOptions } from './getQueryOptions';
-import { getValues } from './getValues';
-import { getQueryToFindUpstreamIds } from './getQueryToFindUpstreamIds';
-
-export class FhirServiceRequest extends FhirResource {
-  static init(options, models) {
-    super.init(
-      {
-        contained: DataTypes.JSONB,
-        identifier: DataTypes.JSONB,
-        status: {
-          type: DataTypes.TEXT,
-          allowNull: false,
-        },
-        intent: {
-          type: DataTypes.TEXT,
-          allowNull: false,
-        },
-        category: DataTypes.JSONB,
-        priority: DataTypes.TEXT,
-        code: DataTypes.JSONB,
-        orderDetail: DataTypes.JSONB,
-        subject: {
-          type: DataTypes.JSONB,
-          allowNull: false,
-        },
-        encounter: DataTypes.JSONB,
-        occurrenceDateTime: DataTypes.TEXT,
-        requester: DataTypes.JSONB,
-        locationCode: DataTypes.JSONB,
-        note: DataTypes.JSONB,
-      },
-      options,
-    );
-
-    this.UpstreamModels = [models.ImagingRequest, models.LabRequest];
-    this.upstreams = [
-      models.ImagingRequest,
-      models.ImagingRequestArea,
-      models.ImagingAreaExternalCode,
-      models.Encounter,
-      models.Facility,
-      models.Location,
-      models.LocationGroup,
-      models.Patient,
-      models.ReferenceData,
-      models.User,
-    ];
-  }
-
-  static CAN_DO = new Set([
-    FHIR_INTERACTIONS.INSTANCE.READ,
-    FHIR_INTERACTIONS.TYPE.SEARCH,
-    FHIR_INTERACTIONS.INTERNAL.MATERIALISE,
-  ]);
-
-  async updateMaterialisation() {
-    const upstream = await this.getUpstream(getQueryOptions(this.sequelize.models));
-    const values = await getValues(upstream, this.sequelize.models);
-    this.set(values);
-  }
-
-  static async queryToFindUpstreamIdsFromTable(upstreamTable, table, id) {
-    return getQueryToFindUpstreamIds(upstreamTable, this.sequelize.models, table, id);
-  }
-
-  static searchParameters() {
-    return {
-      ...super.searchParameters(),
-      identifier: {
-        type: FHIR_SEARCH_PARAMETERS.TOKEN,
-        path: [['identifier', '[]']],
-        tokenType: FHIR_SEARCH_TOKEN_TYPES.VALUE,
-      },
-      category: {
-        type: FHIR_SEARCH_PARAMETERS.TOKEN,
-        path: [['category', '[]', 'coding', '[]']],
-        tokenType: FHIR_SEARCH_TOKEN_TYPES.CODING,
-      },
-      intent: {
-        type: FHIR_SEARCH_PARAMETERS.STRING,
-        path: [['intent']],
-      },
-      occurrence: {
-        type: FHIR_SEARCH_PARAMETERS.DATE,
-        path: [['occurrenceDateTime']],
-      },
-      priority: {
-        type: FHIR_SEARCH_PARAMETERS.STRING,
-        path: [['priority']],
-      },
-      status: {
-        type: FHIR_SEARCH_PARAMETERS.STRING,
-        path: [['status']],
-      },
-    };
-  }
-}
+export * from './model';
